+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [next]
+## [1.4.1]
 - Optimize wildcard search(e.g.: `microsoft*bookmark`), up to 5X+ speed up for cases.
 - Bugfix: active & open files should be search first
 - Optimize: dedup keywords with prefix